pyproject.toml
[project]
name = "fivemap"
version = "0.1.0"
description = "WiFi environment mapper and security audit tool using WiFi Pineapple Mark VII"
requires-python = ">=3.9"
dependencies = [
    "scapy>=2.5.0",
    "paho-mqtt>=1.6.0,<2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "numpy>=1.24.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "boto3>=1.28.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py39"
